news_aggregator/requirements.txt
flask
mysql-connector-python
python-dotenv
requests
schedule
bcrypt
email-validators